refactor: split twitch config and registry helpers
